--- client/landing/stepper/declarative-flow/internals/steps-repository/site-migration-instructions/sidebar/style.scss
@@ -1,23 +1,0 @@
-@import "@automattic/typography/styles/fonts";
-
-.migration-instructions-sidebar {
-	display: flex;
-	flex-direction: column;
-	height: 100%;
-}
-
-.migration-instructions-sidebar__progress {
-	margin-bottom: 2rem;
-}
-
-.migration-instructions-sidebar__title {
-	font-family: $brand-serif;
-	font-size: 2.75rem;
-	line-height: 1.3;
-	margin-bottom: 0.5rem;
-}
-
-.migration-instructions-sidebar__description {
-	font-size: 0.875rem;
-	margin-bottom: 1rem;
-}
