+++ client/landing/stepper/declarative-flow/internals/steps-repository/site-migration-instructions/style.scss
@@ -15,3 +15,9 @@
 	flex: 1;
 	margin-bottom: 2.75rem;
 }
+
+.site-migration-instructions__container-without-preview {
+	max-width: 562px;
+	margin: 60px auto;
+	padding-right: 52px;
+}
